requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
streamlit_pdf_viewer
validators
pyttsx3
pymupdf
pillow
pdf2image
langchain-ollama
